soil_analyzer/android_app/requirements.txt
kivy
kivymd
pandas
numpy
requests
python-dotenv
pillow
buildozer